requirements.txt
discord-py==2.2.2
playwright==1.32.1
